pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ha-hk-citation"
version = "1.0.0"
description = "Home Assistant integration for monitoring Harman Kardon Citation speaker health"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"

[project.urls]
repository = "https://github.com/OWNER/ha-hk-citation"
"Bug Tracker" = "https://github.com/OWNER/ha-hk-citation/issues"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-aiohttp",
    "pytest-cov",
    "pytest-freezer",
    "pytest-socket",
    "pytest-timeout",
    "pytest-homeassistant-custom-component",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["custom_components/hk_citation"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "RUF", "B", "C4", "SIM"]
ignore = ["E501"]
